+++ PSReleaseTools.psm1
@@ -1,5 +1,6 @@
 #load functions
-. $PSScriptRoot\functions.ps1
+. $PSScriptRoot\functions\private.ps1
+. $PSScriptRoot\functions\public.ps1
 
 #configure TLS settings for GitHub
 [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12